sw_engine: --size reduction
basically, sw engine aims for 32bits, it reduces
the bundle size by 1kb.

Need to carefully see any side effect in practice.

+++ src/renderer/sw_engine/tvgSwCommon.h
@@ -33,9 +33,8 @@
 #define SW_ANGLE_2PI (SW_ANGLE_PI << 1)
 #define SW_ANGLE_PI2 (SW_ANGLE_PI >> 1)
 
-using SwCoord = signed long;
-using SwFixed = signed long long;
-
+using SwCoord = int32_t;
+using SwFixed = int64_t;
 
 static inline float TO_FLOAT(SwCoord val)
 {
